refactoring start
