+++ .gitignore
@@ -8,16 +8,11 @@
 *.blg
 *.bbl
 *.synctex.gz
+*.DS_Store
 relaxed.egg-info
 consistent-trees/
 !consistent-trees/read_tree/example.c
-packages/
 .vscode/
-figures/
-!figures/code/
 *temp/
-build/
-data/
-*.DS_Store
-output/30k/
 bin/temp.py
+data/raw_data
