+++ bin/diffmah_fits.py
@@ -4,8 +4,8 @@ from diffmah.fit_mah_helpers import log_mah_mse_loss_and_grads
 from diffmah.utils import jax_adam_wrapper
 from tqdm import tqdm
 
-from relaxed.analysis import get_mah
-from relaxed.analysis import get_t_from_a
+from relaxed.mah import get_mah
+from relaxed.mah import get_t_from_a
 
 
 mah_data = get_mah("m12", "output", cutoff_missing=0.05, cutoff_particle=0.05)
