+++ data/configs/Bolshoi.config
@@ -1,0 +1,9 @@
+[minh]
+
+L = 250
+Epsilon = 1
+Mp = 1.36297e+08
+MinParticles = 50
+NameIndex = /home/imendoza/nbody-relaxed/params/name_index.txt
+TypeIndex = /home/imendoza/nbody-relaxed/params/full_type_index.txt
+
