+++ data/configs/BolshoiP.config
@@ -1,0 +1,9 @@
+[minh]
+
+L = 250
+Epsilon = 1
+Mp = 1.54975e+08
+MinParticles = 50
+NameIndex = name_index.txt
+TypeIndex = full_type_index.txt
+
