+++ data/configs/full_type_index.txt
@@ -1,0 +1,113 @@
+x float32
+y float32
+z float32
+
+vx float32
+vy float32
+vz float32
+
+jx float32
+jy float32
+jz float32
+
+ax float32
+ay float32
+az float32
+
+ax(500c) float32
+ay(500c) float32
+az(500c) float32
+
+gamma_log_vmax_tdyn              float32
+gamma_log_vmax_inst              float32
+log_(vmax/vmax_max(tdyn;tmpeak)) float32
+
+gamma_tdyn   float32
+gamma_100myr float32
+gamma_2tdyn  float32
+gamma_3tdyn  float32
+gamma_4tdyn  float32
+gamma_inst   float32
+gamma_mpeak  float32
+
+snap_num int64
+
+scale            float32
+halfmass_scale   float32
+acc_scale        float32
+m4%_scale        float32
+mpeak_scale      float32
+first_acc_scale  float32
+scale_of_last_mm float32
+
+b_to_a       float32
+b_to_a(500c) float32
+c_to_a       float32
+c_to_a(500c) float32
+
+breadth_first_id                  int64
+depth_first_id                    int64
+desc_id                           int64
+desc_pid                          int64
+descid                            int64
+last_mainleaf_depthfirst_id       int64
+last_progenitor_depthfirst_id     int64
+next_coprogenitor_depthfirst_id   int64
+orig_halo_id                      int64
+tree_root_id                      int64
+
+future_merger_mmp_id  int64
+desc_scale            float32
+time_to_future_merger float32
+
+first_acc_mvir float32
+first_acc_vmax float32
+
+id   int64
+pid  int64
+upid int64
+
+mmp?     int64
+np       int64
+num_prog int64
+phantom  int64
+
+mvir          float32
+mvir_all      float32
+m200b_all     float32
+orig_mvir     float32
+sam_mvir      float32
+sam_m200b     float32
+m200b         float32
+m200c         float32
+m500c         float32
+m2500c        float32
+m_pe_behroozi float32
+m_pe_diemer   float32
+
+macc  float32
+mpeak float32
+
+rvir            float32
+r200b           float32
+rs              float32
+rs_klypin       float32
+rvmax           float32
+halfmass_radius float32
+
+spin         float32
+spin_bullock float32
+
+t/|u| float32
+voff  float32
+xoff  float32
+
+tidal_force      float32
+tidal_force_tdyn float32
+tidal_id         int64
+
+vmax        float32
+vpeak       float32
+vacc        float32
+vmax@mpeak  float32
+vrms        float32
