+++ data/configs/name_index.txt
@@ -1,0 +1,112 @@
+x
+y
+z
+
+vx
+vy
+vz
+
+jx
+jy
+jz
+
+ax a[x]
+ay a[y]
+az a[z]
+
+ax(500c) a[x](500c)
+ay(500c) a[y](500c)
+az(500c) a[z](500c)
+
+gamma_log_vmax_tdyn acc_log_vmax_1*tdyn
+gamma_log_vmax_inst acc_log_vmax_inst
+log_(vmax/vmax_max(tdyn;tmpeak))
+
+gamma_tdyn   acc_rate_1*tdyn acc_rate_tdyn acc_rate_tdyn_tdyn
+gamma_100myr acc_rate_100myr
+gamma_2tdyn  acc_rate_2*tdyn acc_rate_2tdyn_tdyn
+gamma_3tdyn  acc_rate_3*tdyn
+gamma_4tdyn  acc_rate_4*tdyn
+gamma_inst   acc_rate_inst
+gamma_mpeak  acc_rate_mpeak
+
+snap_num snap_idx
+scale
+halfmass_scale
+acc_scale
+m4%_scale
+mpeak_scale
+first_acc_scale
+scale_of_last_mm last_mm
+
+b_to_a
+b_to_a(500c)
+c_to_a
+c_to_a(500c)
+
+breadth_first_id
+depth_first_id
+desc_id
+desc_pid
+desc_scale
+descid
+last_mainleaf_depthfirst_id
+last_progenitor_depthfirst_id
+next_coprogenitor_depthfirst_id
+orig_halo_id
+tree_root_id
+
+future_merger_mmp_id
+time_to_future_merger
+
+first_acc_mvir
+first_acc_vmax
+
+id
+pid
+upid
+
+mmp? mmp
+np
+num_prog
+phantom
+
+mvir     mmvir
+mvir_all mmvir_all
+m200b_all
+sam_mvir sam_mmvir
+sam_m200b
+orig_mvir
+m200b
+m200c
+m500c
+m2500c
+m_pe_behroozi
+m_pe_diemer
+
+macc
+mpeak
+
+rvir rmvir
+r200b
+rs
+rs_klypin
+rvmax
+halfmass_radius
+
+spin
+spin_bullock
+
+t/|u|
+voff
+xoff
+
+tidal_force
+tidal_force_tdyn
+tidal_id
+
+vmax
+vpeak
+vacc
+vmax@mpeak vmax\@mpeak
+vrms
