+++ data/output/m12/z_map.json
@@ -1,0 +1,182 @@
+{
+    "0": 0.07835,
+    "1": 0.09635,
+    "2": 0.10235,
+    "3": 0.10835,
+    "4": 0.11435,
+    "5": 0.12035,
+    "6": 0.13235,
+    "7": 0.13835,
+    "8": 0.14435,
+    "9": 0.15035,
+    "10": 0.15635,
+    "11": 0.16235,
+    "12": 0.16835,
+    "13": 0.17435,
+    "14": 0.18035,
+    "15": 0.18635,
+    "16": 0.19235,
+    "17": 0.19835,
+    "18": 0.20235,
+    "19": 0.20435,
+    "20": 0.21035,
+    "21": 0.21635,
+    "22": 0.22235,
+    "23": 0.22835,
+    "24": 0.23435,
+    "25": 0.24635,
+    "26": 0.25235,
+    "27": 0.25835,
+    "28": 0.26435,
+    "29": 0.27035,
+    "30": 0.27635,
+    "31": 0.28235,
+    "32": 0.28835,
+    "33": 0.29435,
+    "34": 0.30635,
+    "35": 0.31235,
+    "36": 0.31835,
+    "37": 0.32435,
+    "38": 0.33035,
+    "39": 0.33635,
+    "40": 0.34235,
+    "41": 0.34835,
+    "42": 0.35435,
+    "43": 0.36035,
+    "44": 0.36635,
+    "45": 0.37235,
+    "46": 0.37835,
+    "47": 0.38435,
+    "48": 0.39035,
+    "49": 0.39635,
+    "50": 0.40235,
+    "51": 0.40835,
+    "52": 0.41435,
+    "53": 0.42035,
+    "54": 0.42635,
+    "55": 0.43235,
+    "56": 0.43835,
+    "57": 0.44435,
+    "58": 0.45035,
+    "59": 0.45635,
+    "60": 0.46235,
+    "61": 0.46835,
+    "62": 0.47435,
+    "63": 0.48035,
+    "64": 0.48635,
+    "65": 0.49835,
+    "66": 0.50435,
+    "67": 0.51035,
+    "68": 0.51635,
+    "69": 0.52235,
+    "70": 0.52835,
+    "71": 0.53235,
+    "72": 0.53835,
+    "73": 0.54435,
+    "74": 0.55035,
+    "75": 0.55635,
+    "76": 0.56235,
+    "77": 0.56835,
+    "78": 0.57435,
+    "79": 0.58035,
+    "80": 0.58635,
+    "81": 0.59235,
+    "82": 0.59835,
+    "83": 0.60435,
+    "84": 0.61035,
+    "85": 0.61635,
+    "86": 0.62235,
+    "87": 0.62835,
+    "88": 0.63435,
+    "89": 0.64035,
+    "90": 0.64635,
+    "91": 0.65235,
+    "92": 0.65835,
+    "93": 0.66435,
+    "94": 0.67035,
+    "95": 0.67635,
+    "96": 0.68235,
+    "97": 0.68835,
+    "98": 0.69435,
+    "99": 0.70035,
+    "100": 0.70635,
+    "101": 0.71235,
+    "102": 0.71835,
+    "103": 0.72435,
+    "104": 0.73035,
+    "105": 0.73635,
+    "106": 0.74235,
+    "107": 0.74835,
+    "108": 0.75435,
+    "109": 0.76035,
+    "110": 0.76635,
+    "111": 0.77235,
+    "112": 0.77835,
+    "113": 0.78435,
+    "114": 0.79035,
+    "115": 0.79635,
+    "116": 0.80235,
+    "117": 0.80835,
+    "118": 0.81135,
+    "119": 0.81435,
+    "120": 0.81735,
+    "121": 0.82035,
+    "122": 0.82335,
+    "123": 0.82635,
+    "124": 0.82935,
+    "125": 0.83235,
+    "126": 0.83535,
+    "127": 0.83835,
+    "128": 0.84135,
+    "129": 0.84435,
+    "130": 0.84735,
+    "131": 0.85035,
+    "132": 0.85335,
+    "133": 0.85635,
+    "134": 0.85935,
+    "135": 0.86235,
+    "136": 0.86535,
+    "137": 0.86835,
+    "138": 0.87135,
+    "139": 0.87435,
+    "140": 0.87735,
+    "141": 0.88035,
+    "142": 0.88335,
+    "143": 0.88635,
+    "144": 0.88935,
+    "145": 0.89235,
+    "146": 0.89535,
+    "147": 0.89835,
+    "148": 0.90135,
+    "149": 0.90435,
+    "150": 0.90735,
+    "151": 0.91035,
+    "152": 0.91335,
+    "153": 0.91635,
+    "154": 0.91935,
+    "155": 0.92235,
+    "156": 0.92535,
+    "157": 0.92835,
+    "158": 0.93135,
+    "159": 0.93435,
+    "160": 0.93735,
+    "161": 0.94335,
+    "162": 0.94635,
+    "163": 0.94935,
+    "164": 0.95235,
+    "165": 0.95835,
+    "166": 0.96135,
+    "167": 0.96435,
+    "168": 0.96735,
+    "169": 0.97035,
+    "170": 0.97335,
+    "171": 0.97635,
+    "172": 0.97935,
+    "173": 0.98235,
+    "174": 0.98535,
+    "175": 0.98835,
+    "176": 0.99135,
+    "177": 0.99435,
+    "178": 0.99735,
+    "179": 1.00035
+}
