--- minh.md
@@ -1,75 +1,0 @@
-# Available columns
-
-'scale', 
-'id', 
-'desc_scale', 
-'desc_id', 
-'num_prog', 
-'pid', 
-'upid', 
-'desc_pid', 
-'phantom',
- 'sam_mvir',
- 'mvir',
- 'rvir',
- 'rs',
- 'vrms',
- 'mmp?',
- 'scale_of_last_mm',
- 'vmax',
- 'x',
- 'y',
- 'z',
- 'vx',
- 'vy',
- 'vz',
- 'jx',
- 'jy',
- 'jz',
- 'spin',
- 'breadth_first_id',
- 'depth_first_id',
- 'tree_root_id',
- 'orig_halo_id',
- 'snap_num',
- 'next_coprogenitor_depthfirst_id',
- 'last_progenitor_depthfirst_id',
- 'last_mainleaf_depthfirst_id',
- 'rs_klypin',
- 'mvir_all',
- 'm200b',
- 'm200c',
- 'm500c',
- 'm2500c',
- 'xoff',
- 'voff',
- 'spin_bullock',
- 'b_to_a',
- 'c_to_a',
- 'ax',
- 'ay',
- 'az',
- 'b_to_a(500c)',
- 'c_to_a(500c)',
- 'ax(500c)',
- 'ay(500c)',
- 'az(500c)',
- 't/|u|',
- 'm_pe_behroozi',
- 'm_pe_diemer',
- 'macc',
- 'mpeak',
- 'vacc',
- 'vpeak',
- 'halfmass_scale',
- 'gamma_inst',
- 'gamma_100myr',
- 'gamma_tdyn',
- 'gamma_2tdyn',
- 'gamma_mpeak',
- 'mpeak_scale',
- 'acc_scale',
- 'first_acc_scale',
- 'first_acc_mvir',
- 'first_acc_vmax',
- 'vmax@mpeak'
