+++ pyproject.toml
@@ -17,30 +17,29 @@ repository = "https://github.com/ismael-mendoza/nbody-relaxed"
 version = "0.0.1"
 
 [tool.poetry.dependencies]
-astropy = "^4.3.1"
-click = "^8.0.0"
-corner = "^2.2.1"
-findiff = "^0.8.9"
-flake8 = "^4.0.1"
-lmfit = "^1.0.3"
-matplotlib = "^3.5.0"
-numpy = "^1.21.4"
+astropy = ">=4.3.1"
+click = ">=8.0.0"
+corner = ">=2.2.1"
+findiff = ">=0.8.9"
+flake8 = ">=4.0.1"
+lmfit = ">=1.0.3"
+matplotlib = ">=3.5.0"
+numpy = ">=1.21.4"
 pminh = {git = "https://github.com/ismael-mendoza/minnow", branch = "master"}
 python = ">=3.7,<3.11"
-scikit-learn = "^1.0.2 "
-scipy = "^1.7.3"
-tqdm = "^4.63.0"
+scikit-learn = ">=1.0.0"
+scipy = ">=1.7.3"
+tqdm = ">=4.63.0"
 
 [tool.poetry.dev-dependencies]
-Cython = "^0.29.24"
-black = "^22.1.0"
-flake8 = "^4.0.1"
-jupyter-contrib-nbextensions = "^0.5.1"
-jupyterlab = "^3.2.5"
-pre-commit = "^2.16.0"
-pytest = "^6.2.5"
-pytest-cov = "^2.12.1"
-reorder-python-imports = "^2.6.0"
+Cython = ">=0.29.24"
+black = ">=22.1.0"
+flake8 = ">=4.0.1"
+jupyter-contrib-nbextensions = ">=0.5.1"
+pre-commit = ">=2.16.0"
+pytest = ">=6.2.5"
+pytest-cov = ">=2.12.1"
+reorder-python-imports = ">=2.6.0"
 
 [tool.poetry.urls]
 "Bug Tracker" = "https://github.com/ismael-mendoza/nbody-relaxed/issues"
