+++ relaxed/progenitors/progenitor_lines.py
@@ -1,3 +1,4 @@
+"""Module to read main progenitor lines into python objects from text file."""
 import re
 
 from astropy.table import Table
