+++ relaxed/utils.py
@@ -1,0 +1,6 @@
+import numpy as np
+from scipy import stats
+
+
+def get_quantiles(arr):
+    return np.vectorize(lambda x: stats.percentileofscore(arr, x))(arr) / 100.0
